[edited] part 7 of scala code patching It never ends :skull:

+++ src/main/scala/li/cil/oc/client/gui/CustomGuiContainer.scala
@@ -34,7 +34,7 @@ abstract class CustomGuiContainer[C <: AbstractContainerMenu](val inventoryConta
 
   override def windowY = topPos
 
-  override def windowZ = getBlitOffset
+  override def windowZ = getBlitOffset.toFloat
 
   override def isPauseScreen = false
